+++ lib/src/psapi.dart
@@ -4,32 +4,85 @@
 
 // Maps FFI prototypes onto the corresponding Win32 API function calls
 
+// THIS FILE IS GENERATED AUTOMATICALLY AND SHOULD NOT BE EDITED DIRECTLY.
+
+// ignore_for_file: unused_import
+
 import 'dart:ffi';
 
-import 'typedefs.dart';
+import 'package:ffi/ffi.dart';
+
+import 'com/combase.dart';
+import 'structs.dart';
 
 final _psapi = DynamicLibrary.open('psapi.dll');
 
-/// {@category psapi}
-final EnumProcesses = _psapi
-    .lookupFunction<enumProcessesNative, enumProcessesDart>('EnumProcesses');
+// BOOL EnumProcesses(
+//   DWORD   *lpidProcess,
+//   DWORD   cb,
+//   LPDWORD lpcbNeeded
+// );
 
 /// {@category psapi}
-final EnumProcessModules =
-    _psapi.lookupFunction<enumProcessModulesNative, enumProcessModulesDart>(
-        'EnumProcessModules');
+final EnumProcesses = _psapi.lookupFunction<
+    Int32 Function(
+        Pointer<Uint32> lpidProcess, Uint32 cb, Pointer<Uint32> lpcbNeeded),
+    int Function(Pointer<Uint32> lpidProcess, int cb,
+        Pointer<Uint32> lpcbNeeded)>('EnumProcesses');
+
+// BOOL EnumProcessModules(
+//   HANDLE  hProcess,
+//   HMODULE *lphModule,
+//   DWORD   cb,
+//   LPDWORD lpcbNeeded
+// );
 
 /// {@category psapi}
-final EnumProcessModulesEx =
-    _psapi.lookupFunction<enumProcessModulesExNative, enumProcessModulesExDart>(
-        'EnumProcessModulesEx');
+final EnumProcessModules = _psapi.lookupFunction<
+    Int32 Function(IntPtr hProcess, Pointer<IntPtr> lphModule, Uint32 cb,
+        Pointer<Uint32> lpcbNeeded),
+    int Function(int hProcess, Pointer<IntPtr> lphModule, int cb,
+        Pointer<Uint32> lpcbNeeded)>('EnumProcessModules');
+
+// BOOL EnumProcessModulesEx(
+//   HANDLE  hProcess,
+//   HMODULE *lphModule,
+//   DWORD   cb,
+//   LPDWORD lpcbNeeded,
+//   DWORD   dwFilterFlag
+// );
 
 /// {@category psapi}
-final GetModuleBaseName =
-    _psapi.lookupFunction<getModuleBaseNameNative, getModuleBaseNameDart>(
-        'GetModuleBaseNameW');
+final EnumProcessModulesEx = _psapi.lookupFunction<
+    Int32 Function(IntPtr hProcess, Pointer<IntPtr> lphModule, Uint32 cb,
+        Pointer<Uint32> lpcbNeeded, Uint32 dwFilterFlag),
+    int Function(int hProcess, Pointer<IntPtr> lphModule, int cb,
+        Pointer<Uint32> lpcbNeeded, int dwFilterFlag)>('EnumProcessModulesEx');
+
+// DWORD GetModuleBaseNameW(
+//   HANDLE  hProcess,
+//   HMODULE hModule,
+//   LPWSTR  lpBaseName,
+//   DWORD   nSize
+// );
 
 /// {@category psapi}
-final GetModuleFileNameEx =
-    _psapi.lookupFunction<getModuleFileNameExNative, getModuleFileNameExDart>(
-        'GetModuleFileNameExW');
+final GetModuleBaseName = _psapi.lookupFunction<
+    Uint32 Function(IntPtr hProcess, IntPtr hModule, Pointer<Utf16> lpBaseName,
+        Uint32 nSize),
+    int Function(int hProcess, int hModule, Pointer<Utf16> lpBaseName,
+        int nSize)>('GetModuleBaseNameW');
+
+// DWORD GetModuleFileNameExW(
+//   HANDLE  hProcess,
+//   HMODULE hModule,
+//   LPWSTR  lpFilename,
+//   DWORD   nSize
+// );
+
+/// {@category psapi}
+final GetModuleFileNameEx = _psapi.lookupFunction<
+    Uint32 Function(IntPtr hProcess, IntPtr hModule, Pointer<Utf16> lpFilename,
+        Uint32 nSize),
+    int Function(int hProcess, int hModule, Pointer<Utf16> lpFilename,
+        int nSize)>('GetModuleFileNameExW');