+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.7.0
+
+- Changed how the C-style APIs are generated. This should result in far better
+  code smarts in your editor, as well as major improvements to the
+  auto-generated documentation.
+
 ## 1.6.10
 
 - New APIs